Large file fix - Minimum Transaction size 12
Added extra wait time to ensure extra large files with multiple ledgers
finish properly.  before this change 1 out of 4 tests did not complete
properly due to RPC calls failing to return a transaction ID that was
entered on the previous loop.  10 out of 10 tests returned after adding
a 1 second pause between each sendmany call.

Entering a transaction size of less than 12 could cause an infinite loop
and drain a person wallet of funds.  :(   added code to not allow this
setting if the wallet gui is used.