Bug Fix and Web Index changes
Fixed a bug preventing the rendering of some archived images.
Added the ability to recover a partial archive be selecting the  address
file without an associated ledger file and archiving.
added ssi includes to the main index allowing users to add meta tags,
css and header and footer info to all indexed transactions