Implemented Signed Archives: Security enhancement
Their is now a feature that allows users to sign an archive with an
owned address that is than auto verified by any client.    changed safe
file monitoring to use a white list.
Fixed a bug that would have allowed a file of an exact size to bypass
the safe file checks.