Prep for 0.3.0 Release - Profile \ Proof
Added ability to search by file or Hash to Proof code.  With the new
link feature this allows an individual to archive full data on cheap
blockchain and link this to the Hash of the Object of bitcoin  for
lookup.  What good is a hash if the file is deleted ;)

Prep work for Profiles